backend/requirements.txt
Flask
python-dotenv
azure-storage-blob 
azure-cosmos
azure-identity
langchain
openai==0.28.1
pypdf
nltk
azure-cognitiveservices-speech
azure.storage.queue 